+++ include/delfem2/pbd_geo3dtri23.h
@@ -25,7 +25,7 @@ namespace delfem2 {
  */
 void PBD_TriStrain(
     double* aXYZt,
-    unsigned int nXYZ,
+    size_t nXYZ,
     const std::vector<delfem2::CDynTri>& aETri,
     const std::vector<CVec2d>& aVec2);
 
@@ -39,7 +39,7 @@ void PBD_TriStrain(
  */
 void PBD_Bend(
     double* aXYZt,
-    unsigned int nXYZ,
+    size_t nXYZ,
     const std::vector<delfem2::CDynTri>& aETri,
     const std::vector<CVec2d>& aVec2,
     double ratio);
